pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "noctyl"
version = "0.1.0"
description = "Static token usage estimator for multi-agent AI workflows"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6",
]

[project.scripts]
noctyl = "noctyl.cli:main"

[project.optional-dependencies]
dev = ["pytest>=7"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
